wip: adding tests
